+++ src-api/resolvers/Authorization.js
@@ -1,0 +1,11 @@
+export const Query = {
+
+    me: (_, args, { currentPlayer }) => currentPlayer
+
+}
+
+export const Mutation = {
+
+    
+
+}
