+++ src-api/resolvers/index.js
@@ -1,13 +1,16 @@
 import * as connections from './Connections'
+import * as authorization from './Authorization'
 import * as teams from './Teams'
 
 export default {
     Query: {
         ...connections.Query,
+        ...authorization.Query,
         ...teams.Query
     },
     Mutation: {
-        ...teams.Mutation
+        ...teams.Mutation,
+        ...authorization.Mutation
     },
     Player: {
         ...teams.Player
